Draft no classpath.txt
ovation.util.ovation_classloader creates a new URLClassLoader with the
Ovation JAR at the top of the classpath list. It uses existing
classloader as the parent. In other words, we accomplish what we were
doing with classpath.txt, but in an isolated class loader.

Everything works, including DSC creation (by ovation.log), but the
connectMethod.invoke call doesn't return the DSC to matlab.